+++ src/Volo/FrontendBundle/Resources/public/js/app/View/CheckoutDeliveryValidationView.js
@@ -4,7 +4,8 @@ var CheckoutDeliveryValidationView = Backbone.View.extend({
         'focus #postal_index_form_input': '_hideTooltip',
         'click': '_hideTooltip',
         'change #city, #postal_index_form_input': '_geoCodeAndValidateDelivery',
-        "keydown #address_line1, #address_line2": '_geoCodeAndValidateDelivery'
+        "blur #address_line1, #address_line2": '_geoCodeAndValidateDelivery',
+        "keyup #address_line1, #address_line2":'_validateForm'
     },
 
     initialize: function (options) {
@@ -73,6 +74,8 @@ var CheckoutDeliveryValidationView = Backbone.View.extend({
             !this.$("#delivery_information_form_button").attr('disabled');
 
         if (submitAllowed) {
+            this.$('#error_msg_delivery_not_saved').addClass('hide');
+
             this.trigger('submit:successful_before', {
                 deliveryTime: $('#order-delivery-time').val()
             });
@@ -128,6 +131,7 @@ var CheckoutDeliveryValidationView = Backbone.View.extend({
     },
 
     _toggleSubmitButtonDisabled: function(disabledState) {
+        console.log('_toggleSubmitButtonDisabled ', disabledState);
         this.$("#delivery_information_form_button").attr('disabled', disabledState);
     },
 
@@ -140,16 +144,22 @@ var CheckoutDeliveryValidationView = Backbone.View.extend({
         return streetValid && houseNumValid;
     },
 
-    _geocodeAddress: function() {
-        var fullAddress = this.$('#address_line1').val().length > 0 && this.$('#address_line2').val().length > 0 &&
+    isValidForm: function() {
+        return this.$('#address_line1').val().length > 0 && this.$('#address_line2').val().length > 0 &&
             this.$('#postal_index_form_input').val().length > 0 && this.$('#city').val().length > 0;
+    },
 
+    _validateForm: function() {
+        this._toggleSubmitButtonDisabled(!this.isValidForm());
+    },
+
+    _geocodeAddress: function() {
         this._toggleSubmitButtonDisabled(true);
-        if (!fullAddress) {
+        if (!this.isValidForm()) {
             return;
         }
 
-        this.postalCodeGeocodingService.geocode({
+        this.postalCodeGeocodingService.geocodeAddress({
             address: this.$('#address_line1').val() + ' ' + this.$('#address_line2').val() + ', ' + this.$('#postal_index_form_input').val() + ', ' + this.$('#city').val(),
 
             success: function (result) {
@@ -166,9 +176,10 @@ var CheckoutDeliveryValidationView = Backbone.View.extend({
     },
 
     _geocodePostalCode: function (locationData) {
+        console.log('_geocodePostalCode ', this.cid, locationData);
         this._geocodeAddress();
 
-        this.postalCodeGeocodingService.geocode({
+        this.postalCodeGeocodingService.geocodeCenterPostalcode({
             address: locationData.postcode + ", " + locationData.city,
             postalCode: locationData.postcode,
             city: locationData.city,
@@ -179,7 +190,7 @@ var CheckoutDeliveryValidationView = Backbone.View.extend({
 
             error: function (results, status) {
                 var postalCode;
-                if (results.length > 0) {
+                if (results && results.length > 0) {
                     postalCode = _.findWhere(results[0].address_components, {types: ['postal_code']});
                 }
 
@@ -189,10 +200,7 @@ var CheckoutDeliveryValidationView = Backbone.View.extend({
                 } else if (_.get(postalCode, 'long_name') === this.locationModel.get('postcode')) {
                     this._validateDelivery({lat: this.locationModel.get('latitude'), lng: this.locationModel.get('longitude')});
                 } else {
-                    this._validateDelivery({lat: result.lat(), lng: result.lng()});
-                    // if Google can't geo-code it, who are we to stop the user!!!, just consider it valid man :)
-                    this._toggleSubmitButtonDisabled(false);
-
+                    this._toggleSubmitButtonDisabled(!this.isValidForm());
                 }
             }.bind(this)
         });
@@ -216,11 +224,12 @@ var CheckoutDeliveryValidationView = Backbone.View.extend({
                     if (this.$('#postal_index_form_input').data('is-blocking-popup')) {
                         this._hideTooltip();
                     }
+                    this._toggleSubmitButtonDisabled(!this._validateAddressFields());
                 } else {
                     this._showInputPopup(this.$('#postal_index_form_input').data('validation-msg'), true);
+                    this._toggleSubmitButtonDisabled(true);
                 }
 
-                this._toggleSubmitButtonDisabled(!this._validateAddressFields());
             }.bind(this));
     }
 });